big changes

+++ cultivos/views.py
@@ -2,14 +2,13 @@ from django.shortcuts import render, redirect, get_object_or_404
 from django.contrib.auth import login
 from django.contrib.auth.decorators import login_required
 from django.contrib.auth.forms import UserCreationForm, AuthenticationForm
-from .models import Campo, Cultivo, Plagas,PlanificacionCultivo, Pesticidas, Fertilizantes
+from .models import Campo, Cultivo, Plagas, PlanificacionCultivo, Pesticidas, Fertilizantes
 from .forms import CampoForm, CultivoForm, PlagasForm, PlanificacionCultivoForm, PesticidaForm, FertilizanteForm
 
 def home(request):
     return render(request, 'home.html')
 
 def registrarse(request):
-    
     if request.method == 'POST':
         form = UserCreationForm(request.POST)
         if form.is_valid():
@@ -17,74 +16,55 @@ def registrarse(request):
             return redirect('login')
     else:
         form = UserCreationForm()
-    
-    return render(request, 'registrarse.html', {
-        'form': form
-    })
+    return render(request, 'registrarse.html', {'form': form})
 
 def login_vista(request):
     if request.method == 'POST':
         form = AuthenticationForm(request, data=request.POST)
         if form.is_valid():
-            user = form.get_user()  
-            login(request, user)  
-            return redirect('home') 
+            user = form.get_user()
+            login(request, user)
+            return redirect('home')
     else:
-        form = AuthenticationForm() 
-
-    return render(request, 'login.html', {
-        'form': form
-    })
-
-@login_required(login_url='/login/')
-def crear_campo(request):
-    if request.method == 'POST':
-        form = CampoForm(request.POST, user=request.user)
-        if form.is_valid():
-            form.save()
-            return redirect('admin_campos')
-    else:
-        form = CampoForm(user=request.user)
-
-    return render(request, 'crear_campo.html', {
-        'form': form
-    })
+        form = AuthenticationForm()
+    return render(request, 'login.html', {'form': form})
 
 @login_required(login_url='/login/')
 def admin_campos(request):
     editing = False
-    campo_a_editar = None  
-    campos = Campo.objects.filter(usuario=request.user)
+    campo_a_editar = None  # Almacenará el campo a editar, si lo hay
+    campos = Campo.objects.filter(usuario=request.user)  # Solo mostrar campos del usuario actual
 
     # Si es una solicitud POST
     if request.method == 'POST':
         # Eliminar campo
-        if 'eliminar_campo_id' in request.POST:
-            campo_id = request.POST.get('eliminar_campo_id')
+        if 'eliminar_id' in request.POST:
+            campo_id = request.POST.get('eliminar_id')
             campo = Campo.objects.get(id=campo_id, usuario=request.user)
-            campo.delete()
-            return redirect('admin_campos')
+            campo.delete()  # Elimina el campo
+            return redirect('admin_campos')  # Redirige a la misma página después de eliminar
 
-        # Si el campo ya existe 
+        # Si se está editando un campo
         elif 'campo_id' in request.POST:
             campo_id = request.POST.get('campo_id')
             campo_a_editar = Campo.objects.get(id=campo_id, usuario=request.user)
             form = CampoForm(request.POST, instance=campo_a_editar)
             if form.is_valid():
-                form.save()  
-                return redirect('admin_campos')
-            editing = True  
+                form.save()  # Guarda los cambios en el campo editado
+                return redirect('admin_campos')  # Redirige a la misma página después de guardar
 
-        # Si es un campo nuevo (sin campo_id)
+            editing = True  # Activamos el modo edición
+
+        # Si se está creando un nuevo campo
         else:
             form = CampoForm(request.POST)
             if form.is_valid():
                 nuevo_campo = form.save(commit=False)
-                nuevo_campo.usuario = request.user
-                nuevo_campo.save()
-                return redirect('admin_campos')
+                nuevo_campo.usuario = request.user  # Asignamos el campo al usuario actual
+                nuevo_campo.save()  # Guardamos el nuevo campo
+                return redirect('admin_campos')  # Redirige a la misma página después de crear
 
-    # Si es una solicitud GET o estamos en el modo normal de creacion
+    # Si es una solicitud GET o estamos en el modo normal de creación
     else:
         form = CampoForm()
 
@@ -92,104 +72,97 @@ def admin_campos(request):
         'campos': campos,
         'form': form,
         'editing': editing,
-        'campo_a_editar': campo_a_editar,  # Pasamos el campo que se esta editando
+        'campo_a_editar': campo_a_editar,  # Pasamos el campo a editar, si es necesario
     })
+
 
 @login_required(login_url='/login/')
 def admin_cultivos(request):
     editing = False
-    cultivo_a_editar = None  # Variable para almacenar el cultivo que se va a editar
+    cultivo_a_editar = None  # Almacenará el cultivo a editar, si lo hay
+    cultivos = Cultivo.objects.filter(campo__usuario=request.user)  # Solo mostrar cultivos del usuario actual
 
-    # Si es una solicitud POST, verificar si es para eliminar o editar
+    # Si es una solicitud POST
     if request.method == 'POST':
         # Eliminar cultivo
-        if 'eliminar_cultivo_id' in request.POST:
-            cultivo_id = request.POST.get('eliminar_cultivo_id')
-            cultivo = Cultivo.objects.get(id=cultivo_id)
-            cultivo.delete()
-            return redirect('admin_cultivos')
+        if 'eliminar_id' in request.POST:
+            cultivo_id = request.POST.get('eliminar_id')
+            cultivo = Cultivo.objects.get(id=cultivo_id, campo__usuario=request.user)
+            cultivo.delete()  # Elimina el cultivo
+            return redirect('admin_cultivos')  # Redirige a la misma página después de eliminar
 
-        # Editar cultivo
-        elif 'editar_cultivo_id' in request.POST:
-            cultivo_id = request.POST.get('editar_cultivo_id')
-            cultivo_a_editar = Cultivo.objects.get(id=cultivo_id)
-            form = CultivoForm(instance=cultivo_a_editar, user=request.user)
-            editing = True  # Cambiamos el estado a edición
-
-        # Guardar los cambios o crear nuevo cultivo
-        else:
-            if 'cultivo_id' in request.POST:
-                cultivo_a_editar = Cultivo.objects.get(id=request.POST.get('cultivo_id'))
-                form = CultivoForm(request.POST, instance=cultivo_a_editar, user=request.user)
-                editing = True
-            else:
-                form = CultivoForm(request.POST, user=request.user)
-
+        # Si se está editando un cultivo
+        elif 'cultivo_id' in request.POST:
+            cultivo_id = request.POST.get('cultivo_id')
+            cultivo_a_editar = Cultivo.objects.get(id=cultivo_id, campo__usuario=request.user)
+            form = CultivoForm(request.POST, instance=cultivo_a_editar, user=request.user)
             if form.is_valid():
-                form.save()
-                return redirect('admin_cultivos')
+                form.save()  # Guarda los cambios en el cultivo editado
+                return redirect('admin_cultivos')  # Redirige a la misma página después de guardar
+
+            editing = True  # Activamos el modo edición
+
+        # Si se está creando un nuevo cultivo
+        else:
+            form = CultivoForm(request.POST, user=request.user)
+            if form.is_valid():
+                nuevo_cultivo = form.save(commit=False)
+                nuevo_cultivo.save()  # Guardamos el nuevo cultivo
+                return redirect('admin_cultivos')  # Redirige a la misma página después de crear
 
     # Si es una solicitud GET o estamos en el modo normal de creación
-    if not cultivo_a_editar:
+    else:
         form = CultivoForm(user=request.user)
 
-    # Obtener los cultivos del usuario actual
-    campos_usuario = Campo.objects.filter(usuario=request.user)
-    cultivos = Cultivo.objects.filter(campo__in=campos_usuario)
-
     return render(request, 'admin_cultivos.html', {
-        'form': form,
         'cultivos': cultivos,
+        'form': form,
         'editing': editing,
-        'cultivo_a_editar': cultivo_a_editar,
+        'cultivo_a_editar': cultivo_a_editar,  # Pasamos el cultivo a editar, si es necesario
     })
 
 @login_required(login_url='/login/')
 def admin_plagas(request):
     editing = False
     plaga_a_editar = None
+    # Filtrar las plagas del usuario, pero por campo, no por cultivo
+    plagas = Plagas.objects.filter(cultivo__campo__usuario=request.user)  # Filtrar las plagas por campo
 
-    # Si es una solicitud POST, verificar si es para eliminar o editar
+    # Si la solicitud es POST
     if request.method == 'POST':
         # Eliminar plaga
         if 'eliminar_plaga_id' in request.POST:
             plaga_id = request.POST.get('eliminar_plaga_id')
-            plaga = Plagas.objects.get(id=plaga_id)
+            plaga = Plagas.objects.get(id=plaga_id, cultivo__campo__usuario=request.user)
             plaga.delete()
             return redirect('admin_plagas')
 
         # Editar plaga
-        elif 'editar_plaga_id' in request.POST:
-            plaga_id = request.POST.get('editar_plaga_id')
-            plaga_a_editar = Plagas.objects.get(id=plaga_id)
-            form = PlagasForm(instance=plaga_a_editar)
-            editing = True  # Cambiamos el estado a edición
-
-        # Guardar los cambios o crear nueva plaga
-        else:
-            if 'plaga_id' in request.POST:  # Si existe un ID de plaga en el formulario (modo edición)
-                plaga_a_editar = Plagas.objects.get(id=request.POST.get('plaga_id'))
-                form = PlagasForm(request.POST, instance=plaga_a_editar)
-                editing = True
-            else:  # Si no, estamos creando una nueva plaga
-                form = PlagasForm(request.POST)
-
+        elif 'plaga_id' in request.POST:
+            plaga_id = request.POST.get('plaga_id')
+            plaga_a_editar = Plagas.objects.get(id=plaga_id, cultivo__campo__usuario=request.user)
+            form = PlagasForm(request.POST, instance=plaga_a_editar, user=request.user)
             if form.is_valid():
                 form.save()
                 return redirect('admin_plagas')
+            editing = True
 
-    # Si es una solicitud GET o estamos en el modo normal de creación
-    if not plaga_a_editar:
-        form = PlagasForm()
+        # Crear nueva plaga
+        else:
+            form = PlagasForm(request.POST, user=request.user)
+            if form.is_valid():
+                nueva_plaga = form.save(commit=False)
+                nueva_plaga.save()
+                return redirect('admin_plagas')
 
-    # Obtener todas las plagas
-    plagas = Plagas.objects.all()
+    else:
+        form = PlagasForm(user=request.user)
 
     return render(request, 'admin_plagas.html', {
-        'form': form,
         'plagas': plagas,
+        'form': form,
         'editing': editing,
-        'plaga_a_editar': plaga_a_editar,
+        'plaga_a_editar': plaga_a_editar,  # Pasamos la plaga si estamos editando
     })
 
 
@@ -197,6 +170,9 @@ def admin_plagas(request):
 def admin_planificaciones(request):
     editing = False
     planificacion_a_editar = None
+
+    # Obtener todas las planificaciones del usuario
+    planificaciones = PlanificacionCultivo.objects.filter(campo__usuario=request.user)
 
     if request.method == 'POST':
         # Eliminar planificacion
@@ -229,12 +205,9 @@ def admin_planificaciones(request):
     if not planificacion_a_editar:
         form = PlanificacionCultivoForm(user=request.user)
 
-    # Obtener las planificaciones del usuario actual
-    planificaciones = PlanificacionCultivo.objects.filter(campo__usuario=request.user)
-
     return render(request, 'admin_planificaciones.html', {
         'form': form,
-        'planificaciones': planificaciones,
+        'planificaciones': planificaciones,  # Enviar planificaciones a la plantilla
         'editing': editing,
         'planificacion_a_editar': planificacion_a_editar,
     })
@@ -244,7 +217,6 @@ def admin_pesticidas(request):
     editing = False
     pesticida_a_editar = None
 
-    # Si es una solicitud POST, verificar si es para eliminar o editar
     if request.method == 'POST':
         # Eliminar pesticida
         if 'eliminar_pesticida_id' in request.POST:
@@ -257,28 +229,28 @@ def admin_pesticidas(request):
         elif 'editar_pesticida_id' in request.POST:
             pesticida_id = request.POST.get('editar_pesticida_id')
             pesticida_a_editar = Pesticidas.objects.get(id=pesticida_id)
-            form = PesticidaForm(instance=pesticida_a_editar)
+            form = PesticidaForm(instance=pesticida_a_editar, user=request.user)
             editing = True
 
-        # Guardar cambios o crear nuevo pesticida
         else:
+            # Si no es una solicitud de edición, se crea o actualiza el pesticida
             if 'pesticida_id' in request.POST:
                 pesticida_a_editar = Pesticidas.objects.get(id=request.POST.get('pesticida_id'))
-                form = PesticidaForm(request.POST, instance=pesticida_a_editar)
+                form = PesticidaForm(request.POST, instance=pesticida_a_editar, user=request.user)
                 editing = True
             else:
-                form = PesticidaForm(request.POST)
+                form = PesticidaForm(request.POST, user=request.user)
 
             if form.is_valid():
                 form.save()
                 return redirect('admin_pesticidas')
 
-    # Si es una solicitud GET o estamos en el modo normal de creación
+    # Si es una solicitud GET o estamos en el modo de crear
     if not pesticida_a_editar:
-        form = PesticidaForm()
+        form = PesticidaForm(user=request.user)
 
-    # Obtener los pesticidas
-    pesticidas = Pesticidas.objects.all()
+    # Obtener los pesticidas del usuario actual, filtrados por campo
+    pesticidas = Pesticidas.objects.filter(campo__usuario=request.user)
 
     return render(request, 'admin_pesticidas.html', {
         'form': form,
@@ -297,25 +269,31 @@ def admin_fertilizantes(request):
         # Eliminar fertilizante
         if 'eliminar_fertilizante_id' in request.POST:
             fertilizante_id = request.POST.get('eliminar_fertilizante_id')
-            fertilizante = Fertilizantes.objects.get(id=fertilizante_id)
-            fertilizante.delete()
+            fertilizante = get_object_or_404(Fertilizantes, id=fertilizante_id)
+            # Asegurarse que el fertilizante pertenece al campo del usuario
+            if fertilizante.campo.usuario == request.user:
+                fertilizante.delete()
             return redirect('admin_fertilizantes')
 
         # Editar fertilizante
         elif 'editar_fertilizante_id' in request.POST:
             fertilizante_id = request.POST.get('editar_fertilizante_id')
-            fertilizante_a_editar = Fertilizantes.objects.get(id=fertilizante_id)
-            form = FertilizanteForm(instance=fertilizante_a_editar)
-            editing = True
+            fertilizante_a_editar = get_object_or_404(Fertilizantes, id=fertilizante_id)
+            # Asegurarse que el fertilizante pertenece al campo del usuario
+            if fertilizante_a_editar.campo.usuario == request.user:
+                form = FertilizanteForm(instance=fertilizante_a_editar, user=request.user)
+                editing = True
+            else:
+                return redirect('admin_fertilizantes')
 
         # Guardar cambios o crear nuevo fertilizante
         else:
             if 'fertilizante_id' in request.POST:
-                fertilizante_a_editar = Fertilizantes.objects.get(id=request.POST.get('fertilizante_id'))
-                form = FertilizanteForm(request.POST, instance=fertilizante_a_editar)
+                fertilizante_a_editar = get_object_or_404(Fertilizantes, id=request.POST.get('fertilizante_id'))
+                form = FertilizanteForm(request.POST, instance=fertilizante_a_editar, user=request.user)
                 editing = True
             else:
-                form = FertilizanteForm(request.POST)
+                form = FertilizanteForm(request.POST, user=request.user)
 
             if form.is_valid():
                 form.save()
@@ -323,10 +301,10 @@ def admin_fertilizantes(request):
 
     # Si es una solicitud GET o estamos en el modo normal de creación
     if not fertilizante_a_editar:
-        form = FertilizanteForm()
+        form = FertilizanteForm(user=request.user)
 
-    # Obtener los fertilizantes
-    fertilizantes = Fertilizantes.objects.all()
+    # Obtener los fertilizantes que pertenecen al usuario
+    fertilizantes = Fertilizantes.objects.filter(campo__usuario=request.user)
 
     return render(request, 'admin_fertilizantes.html', {
         'form': form,
@@ -334,6 +312,3 @@ def admin_fertilizantes(request):
         'editing': editing,
         'fertilizante_a_editar': fertilizante_a_editar,
     })
-
-
-
